Endpoints split in several files according to its name.
