+++ api/comments.py
@@ -1,0 +1,10 @@
+# System imports
+
+# Third-party imports
+from flask import Blueprint
+from flask import jsonify
+
+# Local source tree imports
+import config
+
+con = config.CON
