+++ api/napps.py
@@ -1,17 +1,20 @@
 # System imports
 
 # Third-party imports
-from flask import Flask
+from flask import Blueprint
 from flask import jsonify
 
 # Local source tree imports
 import config
 
-app = Flask(__name__)
+# Flask Blueprints
+napps_api = Blueprint('napps_api', __name__)
+napp_api = Blueprint('napp_api', __name__)
+
 con = config.CON
 
 
-@app.route('/apps/', methods=['GET'])
+@napps_api.route('/apps/', methods=['GET'])
 def get_apps():
     app_dict = {"app": {}}
     app_names = con.smembers("apps")
@@ -35,7 +38,7 @@ def get_apps():
     return jsonify(app_dict)
 
 
-@app.route('/apps/<name>', methods=['GET'])
+@napp_api.route('/apps/<name>', methods=['GET'])
 def get_app(name):
     app_dict = {"app": {}}
 
@@ -52,44 +55,3 @@ def get_app(name):
                                               ("app:"+name, "versions")))
 
     return jsonify(app_dict)
-
-
-@app.route('/authors', methods=['GET'])
-def get_authors():
-
-    authors_dict = {"author": {}}
-    authors_names = con.smembers("authors")
-
-    for author_name in authors_names:
-        authors_dict["author"][author_name] = con.hgetall(author_name)
-        authors_dict["author"][author_name]["apps"] = list(con.smembers
-                                                           (con.hget
-                                                            (author_name,
-                                                             "apps")))
-
-        authors_dict["author"][author_name]["comments"] = list(con.smembers
-                                                               (con.hget
-                                                                (author_name,
-                                                                 "comments")))
-
-    return jsonify(authors_dict)
-
-
-@app.route('/authors/<name>', methods=['GET'])
-def get_author(name):
-
-    author_dict = {"author": {}}
-
-    author_dict["author"][name] = con.hgetall(name)
-    author_dict["author"][name]["apps"] = list(con.smembers(con.hget
-                                                            ("author:"+name,
-                                                             "apps")))
-    author_dict["author"][name]["comments"] = list(con.smembers
-                                                   (con.hget
-                                                    ("author:"+name,
-                                                        "comments")))
-    return jsonify(author_dict)
-
-
-if __name__ == '__main__':
-    app.run(debug=True)
